+++ apps/extension/src/hooks/useCraftableItemPage.ts
@@ -2,9 +2,8 @@ import * as React from 'react';
 import { useItemsFetcher } from './useItemsFetcher';
 
 export function useCraftableItemPage(reagentItemIds: number[]) {
-  const id = React.useId();
   const [craftAmount, setCraftAmount] = React.useState(1);
-  const items = useItemsFetcher(id, reagentItemIds);
+  const items = useItemsFetcher(reagentItemIds);
 
   return {
     craftAmount,