+++ apps/extension/src/hooks/useItemsFetcher.ts
@@ -1,56 +1,46 @@
 import * as i from 'types';
-import { useQuery, useQueryClient } from '@tanstack/react-query';
+import { useQueries } from '@tanstack/react-query';
+
 import asyncStorage from 'utils/asyncStorage';
 import { fetchItemFromAPI } from 'src/queries/item';
+
 import useUser from './useUser';
+import { useWowhead } from './useWowhead';
+import { useAuctionHouse } from './useAuctionHouse';
 
-export function useItemsFetcher(id: string | number | undefined, itemIds: number[]) {
-  const queryClient = useQueryClient();
+export function useItemsFetcher(itemIds: number[]) {
   const user = useUser();
+  const { version } = useWowhead();
+  const auctionHouseId = useAuctionHouse();
 
-  return useQuery({
-    queryKey: ['items', id],
-    enabled: !!id && !!user.realm,
-    queryFn: async () => {
-      // Check browser storage if item is stored
-      const cachedItems = await Promise.all(
-        itemIds.map((itemId) => {
-          return asyncStorage.getItem([user.realm!.auctionHouseId, itemId]);
-        }),
-      );
-
-      const itemsIdsToFetch = cachedItems
-        .map((item, index) => {
-          if (item) {
-            return null;
-          }
-
-          return itemIds[index];
-        })
-        .filter(Boolean) as number[];
-
-      const result = await Promise.all(
-        itemsIdsToFetch.map((itemId) => {
-          if (!user.realm || !user.faction) {
-            return;
-          }
-
-          return fetchItemFromAPI(itemId, user.realm.auctionHouseId);
-        }),
-      );
-
-      const allItems = [...cachedItems, ...result];
-
-      for (const item of allItems) {
-        if (!item || !user.realm || typeof item === 'string') {
-          continue;
+  return useQueries({
+    queries: itemIds.map((itemId) => ({
+      enabled: !!user.realm,
+      queryKey: ['item', auctionHouseId, itemId],
+      placeholderData: {
+        itemId,
+      } as i.CachedItemDataClassic,
+      queryFn: async () => {
+        if (!user.realm || !user.faction) {
+          return;
         }
 
-        const itemQueryKey: i.ItemQueryKey = [user.realm.auctionHouseId, item.itemId];
-        queryClient.setQueryData<i.CachedItemDataClassic>(['item', ...itemQueryKey], item);
-      }
+        // Check browser storage if item is stored
+        const cachedItem = await asyncStorage.getItem([user.realm!.auctionHouseId, itemId]);
+        if (cachedItem) {
+          return cachedItem;
+        }
 
-      return allItems.filter((item): item is i.CachedItemDataClassic => !!item);
-    },
+        const fetchedItem = await fetchItemFromAPI(itemId, user.realm.auctionHouseId, version);
+
+        if (fetchedItem === 'NOT_FOUND' || fetchedItem === undefined) {
+          return {
+            name: 'Not found',
+            itemId,
+          } as i.CachedItemDataClassic;
+        }
+        return fetchedItem;
+      },
+    })),
   });
 }