+++ apps/extension/src/hooks/useWowhead.ts
@@ -3,14 +3,14 @@ import * as i from 'types';
 export function useWowhead() {
   let version: i.GameVersion = 'classic';
   if (window.location.href.includes('wowhead.com/classic')) {
-    version = 'era';
+    version = 'seasonal';
   }
   const wowheadBaseUrl =
-    version === 'era' ? 'https://wowhead.com/classic' : 'https://wowhead.com/wotlk';
+    version === 'seasonal' ? 'https://wowhead.com/classic' : 'https://wowhead.com/cata';
 
   return {
-    isEra: version === 'era',
-    isWotlk: version === 'classic',
+    isEra: version === 'seasonal',
+    isClassic: version === 'classic',
     version,
     wowheadBaseUrl,
   };