+++ apps/extension/src/modules/main/tooltip/TooltipBody.tsx
@@ -20,7 +20,8 @@ export const TooltipBody: React.FC<Props> = (props) => {
       return 'Unknown';
     }
 
-    const activeVersion = user.isActive?.[version] || version;
+    const _version = version === 'seasonal' ? 'era' : 'classic'; // Used to be 'era' so for backwards compatibility we change it from 'seasonal' to 'era'
+    const activeVersion = user.isActive?.[_version] || version;
     const serverName = user.realms?.[activeVersion];
     const region = user.region?.toUpperCase();