+++ apps/server/src/types/index.ts
@@ -1,4 +1,4 @@
-export * from './utils/blizzard/types';
+export * from '../utils/blizzard/types';
 
 export namespace NexusHub {
   interface Tooltip {