+++ apps/server/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @auctionoton/server
 
+## 1.2.0
+
+### Minor Changes
+
+- 2a9407d: - Add version to item API
+  - Speed up item API by fetching AH in the background
+  - Fix switching servers between wowhead.com/classic and wowhead.com/cata
+
+## 1.1.0
+
+### Minor Changes
+
+- 3bf82bf: Added queue to fetching auction house
+
 ## 1.0.0
 
 ### Major Changes